+++ CHANGES.rst
@@ -8,6 +8,8 @@ Changes
 - Enhanced support for date and time fields in WFS client.
 - Chrome 118+, Safari 17+, Edge 116+ or Firefox 115+ is required.
 - Enhanced support for JSON output format in the WMS services.
+- Enhanced resource deletion confirmation with details of resources to delete.
+
 
 4.9.0
 -----